Merged main to dimitry

+++ pizza12/src/main/java/fr/eni/pizza12/bo/ProductEntity.java
@@ -12,8 +12,8 @@ public class ProductEntity {
   private boolean productActive;
   private CategoryEntity category;
   private ProductTypeEntity productType;
-    private int productQuantity;
-    private int productOrderId;
+  private int productQuantity;
+  private int productOrderId;
 
   public ProductEntity() {
   }
@@ -34,20 +34,20 @@ public class ProductEntity {
     this(productId, productName, productPrice, productActive, category, null);
   }
 
-    public ProductEntity(int productId, String productName, int productQuantity, int productPrice,
-            boolean productActive,
-            CategoryEntity category) {
-        this(productId, productName, productPrice, productActive, category, null);
-        this.productQuantity = productQuantity;
-    }
+  public ProductEntity(int productId, String productName, int productQuantity, BigDecimal productPrice,
+      boolean productActive,
+      CategoryEntity category) {
+    this(productId, productName, productPrice, productActive, category, null);
+    this.productQuantity = productQuantity;
+  }
 
-    public ProductEntity(int productId, String productName, int productQuantity, int productPrice,
-            boolean productActive,
-            CategoryEntity category, int productOrderId) {
-        this(productId, productName, productPrice, productActive, category, null);
-        this.productQuantity = productQuantity;
-        this.productOrderId = productOrderId;
-    }
+  public ProductEntity(int productId, String productName, int productQuantity, BigDecimal productPrice,
+      boolean productActive,
+      CategoryEntity category, int productOrderId) {
+    this(productId, productName, productPrice, productActive, category, null);
+    this.productQuantity = productQuantity;
+    this.productOrderId = productOrderId;
+  }
 
   public int getProductId() {
     return this.productId;
@@ -101,21 +101,21 @@ public class ProductEntity {
     this.productType = productType;
   }
 
-    public int getProductQuantity() {
-        return this.productQuantity;
-    }
+  public int getProductQuantity() {
+    return this.productQuantity;
+  }
 
-    public void setProductQuantity(int productQuantity) {
-        this.productQuantity = productQuantity;
-    }
+  public void setProductQuantity(int productQuantity) {
+    this.productQuantity = productQuantity;
+  }
 
-    public int getProductOrderId() {
-        return this.productOrderId;
-    }
+  public int getProductOrderId() {
+    return this.productOrderId;
+  }
 
-    public void setProductOrderId(int productOrderId) {
-        this.productOrderId = productOrderId;
-    }
+  public void setProductOrderId(int productOrderId) {
+    this.productOrderId = productOrderId;
+  }
 
   @Override
   public String toString() {
